Pequenos Ajustes
- Troca do shipping_tax para float
- Valor de frete funcional na página do produto

+++ src/database/migrations/20211007010733_add_shipping_tax_to_store.js
@@ -1,6 +1,6 @@
 exports.up = async function (knex) {
   await knex.schema.table('Store', (table) => {
-    table.integer('shipping_tax').notNullable().defaultTo(0);
+    table.float('shipping_tax').notNullable().defaultTo(0);
   });
   await knex.schema.raw(`
         ALTER TABLE "Store"
